requirements.txt
pypdfium2==4.30.1
python-telegram-bot==21.10
